requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.1
dateparser==1.1.1
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.4
mutagen==1.45.1
outcome==1.2.0
pefile==2022.5.30
Pillow==9.2.0
pycparser==2.21
pycryptodomex==3.15.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
regex==2022.3.2
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
webdriver-manager==3.8.3
websockets==10.3
wsproto==1.2.0
yt-dlp==2022.9.1
